backend/InvestmentFilingETL/requirements.txt
requests
beautifulsoup4
PyMuPDF
pdfminer.six
feedparser
psycopg2-binary
